Module sorbus::green

The green tree is an immutable, persistent, atomically reference counted tree.

Structs

Builder
Construction cache for green tree elements.

Checkpoint
Checkpoint for maybe wrapping a node. See TreeBuilder::checkpoint.

Children
Children elements of a node in the immutable green tree.

ChildrenWithOffsets
Children elements of a node in the immutable green tree, with text offsets from the parent node.

Node
A nonleaf node in the immutable green tree.

Token
A leaf token in the immutable green tree.

TreeBuilder
Top-down builder context for a green tree.