Struct sorbus::green::TreeBuilder

pub struct TreeBuilder { /* fields omitted */ }

Top-down builder context for a green tree.

Implementations

impl TreeBuilder

pub fn new() -> Self

Create a new builder.

pub fn new_with(cache: Builder) -> Self

Create a new builder, reusing a Builder cache.

pub fn builder(&mut self) -> &mut Builder

The Builder used to create and deduplicate nodes.

pub fn add(
&mut self,
element: impl Into<NodeOrToken<Arc<Node>, Arc<Token>>>
) -> &mut Self

Add an element to the current branch.

pub fn token(&mut self, kind: Kind, text: &str) -> &mut Self

Add a new token to the current branch.

pub fn node<I, R>(&mut self, kind: Kind, children: I) -> &mut Self where
I: IntoIterator,
I::Item: Into<NodeOrToken<Arc<Node>, Arc<Token>>>,
I::IntoIter: ExactSizeIterator + AsRef<[R]>,
&'a R: Into<NodeOrToken<&'a Node, &'a Token>>,

Add a new node to the current branch.

pub fn start_node(&mut self, kind: Kind) -> &mut Self

Start a new child node and make it the current branch.

pub fn finish_node(&mut self) -> &mut Self

Finish the current branch and restore its parent as current.

pub fn checkpoint(&self) -> Checkpoint

Prepare for maybe wrapping the next node.

To potentially wrap elements into a node, first create a checkpoint, add some items that might be wrapped, then maybe call start_node_at. Don't forget to still call finish_node for the newly started node!

Examples

Checkpoints can be used to implement pratt parsing:

fn parse_expr(b: &mut TreeBuilder, bind: f32, tts: &mut Peekable<impl Iterator<Item=Token>>) {
    let start = b.checkpoint();
    // just assume the tokens are correct for the example
    let first_token = tts.next().unwrap();
    b.token(kind_of(first_token), text_of(first_token));
    loop {
        let power = match tts.peek() {
            None => break,
            Some(op) => binding_power(op),
        };
        if power < bind { break; }
        let op_token = tts.next().unwrap();
        b.token(kind_of(op_token), text_of(op_token));
        parse_expr(&mut *b, power, &mut *tts);
        b.start_node_at(start, EXPR).finish_node();
    }
}

let tokens = vec![ATOM, MUL, ATOM, PLUS, ATOM, MUL, ATOM]; // atom*atom+atom*atom
let expected_tree = builder
    .start_node(EXPR)
        .start_node(EXPR)
            .token(ATOM, "atom")
            .token(MUL, "*")
            .token(ATOM, "atom")
        .finish_node()
        .token(PLUS, "+")
        .start_node(EXPR)
            .token(ATOM, "atom")
            .token(MUL, "*")
            .token(ATOM, "atom")
        .finish_node()
    .finish_node()
    .finish();
parse_expr(&mut builder, 0.0, &mut tokens.into_iter().peekable());
let parsed_tree = builder.finish();
assert_eq!(parsed_tree, expected_tree);

pub fn start_node_at(&mut self, Checkpoint: Checkpoint, kind: Kind) -> &mut Self

Wrap the elements added after checkpoint in a new node, and make the new node the current branch.

pub fn finish_node_at(&mut self, Checkpoint: Checkpoint) -> &mut Self

Finish the current branch up to a given checkpoint, and restore its parent as current.

Any nodes after the used checkpoint will be shifted from the current branch to its parent, after the newly finished node.

Prefer using regular finish_node and delaying adding branches when possible, as its operations on the underlying buffer are marginally more efficient and involve less moving of elements.

pub fn finish(&mut self) -> Arc<Node>

Complete the current tree building.

This TreeBuilder is reset and can be used to build a new tree.

Panics

Panics if more nodes have been started than finished, or the current branch has more than one element.

pub fn recycle(self) -> Builder

Destroy this tree builder and recycle its build cache.